requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
albumentations>=1.3.0
matplotlib>=3.5.0
pandas>=1.3.0
pyyaml>=6.0
ultralytics>=8.0.0  # Required for YOLOv8
ensemble-boxes>=1.0.9 # Required for Weighted Boxes Fusion
tqdm>=4.65.0
scikit-learn>=1.0.0
seaborn
jupyter
ipykernel